backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
pillow>=10.0.0
open_clip_torch>=2.20.0
pdf2image>=1.16.0
openai>=1.0.0
